Create model for insurance
Deciding whether to include gender in the insurance model is a balance between making accurate predictions and being fair. If adding gender helps the model work better without being unfair or biased, it might be a good idea. But if it introduces problems or treats people unfairly based on gender, it's important to think carefully. While a more accurate model can make the company more money, it's crucial to make sure the way we get there is fair and follows the rules.